+++ andes/routines/pflow.py
@@ -8,7 +8,7 @@ from collections import OrderedDict
 from andes.utils.misc import elapsed
 from andes.routines.base import BaseRoutine
 from andes.variables.report import Report
-from andes.shared import np, matrix, sparse, newton_krylov, IP_ADD
+from andes.shared import np, matrix, sparse, newton_krylov
 
 logger = logging.getLogger(__name__)
 
@@ -76,7 +76,7 @@ class PFlow(BaseRoutine):
         self.system.init(self.models, routine='pflow')
         logger.info('Power flow initialized.')
 
-        # force precompile if numba is on - improves timing accuracy
+        # force compile if numba is on - improves timing accuracy
         if system.config.numba:
             system.f_update(self.models)
             system.g_update(self.models)
@@ -149,25 +149,17 @@ class PFlow(BaseRoutine):
         """
         Output a summary for the PFlow routine.
         """
-        ipadd_status = 'Standard (ipadd not available)'
 
         # extract package name, `kvxopt` or `kvxopt`
         sp_module = sparse.__module__
         if '.' in sp_module:
             sp_module = sp_module.split('.')[0]
 
-        if IP_ADD:
-            if self.system.config.ipadd:
-                ipadd_status = f'Fast in-place ({sp_module})'
-            else:
-                ipadd_status = 'Standard (ipadd disabled in config)'
-
         out = list()
         out.append('')
         out.append('-> Power flow calculation')
         out.append(f'{"Sparse solver":>16s}: {self.solver.sparselib.upper()}')
         out.append(f'{"Solution method":>16s}: {self.config.method} method')
-        out.append(f'{"Sparse addition":>16s}: {ipadd_status}')
         out_str = '\n'.join(out)
         logger.info(out_str)