+++ andes/utils/__init__.py
@@ -1,3 +1,4 @@
 from andes.utils import paths           # NOQA
 from andes.utils import sympy           # NOQA
 from andes.utils.paths import get_case  # NOQA
+from andes.utils import snapshot        # NOQA