+++ andes/variables/dae.py
@@ -146,6 +146,9 @@ class DAETimeSeries:
 
         return super().__getattribute__(attr)
 
+    def __getstate__(self):
+        return self.__dict__
+
 
 class DAE:
     r"""
@@ -202,7 +205,7 @@ class DAE:
 
     def __init__(self, system):
         self.system = system
-        self.t = np.array(0)
+        self.t = np.array(0, dtype=float)
         self.ts = DAETimeSeries(self)
 
         self.m, self.n, self.o = 0, 0, 0