+++ docs/source/release-notes.rst
@@ -9,6 +9,18 @@ The APIs before v3.0.0 are in beta and may change without prior notice.
 v1.5 Notes
 ----------
 
+v1.5.2 (2021-10-27)
+```````````````````
+- Removed ``CVXOPT`` dependency.
+- Removed ``__zeros`` and ``__ones`` as they are no longer needed.
+
+- Added ``andes prep -c`` to precompile the generated code.
+- Added utility functions for saving and loading system snapshots.
+  See ``andes/utils/snapshot.py``.
+
+- Compiled numba code is always cached.
+- Bug fixes.
+
 v1.5.1 (2021-10-23)
 ```````````````````
 - Restored compatibility with SymPy 1.6.